refactor: not render header and footer repeatedly
