+++ src/styles/global.css
@@ -21,6 +21,11 @@ p {
     margin: 0;
 }
 
+.display-none,
+.footer.display-none {
+    display: none;
+}
+
 .header-nav {
     position: fixed;
     top: 0;
@@ -120,10 +125,6 @@ p {
     align-items: center;
     gap: 10px;
     cursor: default;
-}
-
-.display-none {
-    display: none;
 }
 
 .active-btn,
